Keys System

+++ app/Http/Controllers/Manager/KeyController.php
@@ -5,8 +5,11 @@ namespace App\Http\Controllers\Manager;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Key;
+use Illuminate\Support\Facades\Validator;
+use App\Traits\MyFunctions;
 class KeyController extends Controller
 {
+    use MyFunctions;
     public function __construct()
     {
         $this->middleware('assign.guard:manager');
@@ -23,34 +26,52 @@ class KeyController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'key' => 'bail|required|string|unique:keys,key|max:30'
         ]);
+
+        if ($validator->fails()) {
+            $response = $this->RespError($validator->errors());
+            return response()->json($response);
+                }
         $key = $request->input('key');
         $create = Key::create([
         'key' => $key
         ]);
-        if(!$create->exists()){
+        if(!$create->exists){
             App::abort(500, 'Error');
         }else{
-            return redirect()->back()->with(['success' => 'تم تسجيل المفتاح بنجاح']);
+            $response = $this->RespSuccess('تم إنشاء المفتاح بنجاح');
+            return response()->json($response);
 
         }
 
     }
 
 
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        $key = Key::where('id',$id)->first();
+        $validator = Validator::make($request->all(), [
+            'id' => 'bail|required|numeric'
+        ]);
 
-        if(!$key)
-        return redirect()->back()->withErrors(['Error' => 'لم يتم العثور على المفتاح']);
+        if ($validator->fails()) {
+            $response = $this->RespError($validator->errors());
+            return response()->json($response);
+                }
+        $id = $request->input('id');
 
+        $key = Key::find($id);
+
+        if(!$key){
+        $response = $this->RespError(['Error' => ['لم يتم العثور على المفتاح']]);
+        return response()->json($response);
+        }
 
         $key->delete();
+        $response = $this->RespSuccess('تم حذف المفتاح بنجاح');
+        return response()->json($response);
 
-        return redirect()->route('KeysCreate')->with(['success' => 'تم حذف المفتاح بنجاح']);
     }
 
 }
